% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{Blues}
\alias{Blues}
\title{Create a Blues palette}
\usage{
Blues(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a Blues palette.
}
\details{
Blues
}
\examples{
Blues(9)
ggPalette(Blues(9))
}
