% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{BrBG}
\alias{BrBG}
\title{Create a BrBG palette}
\usage{
BrBG(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a BrBG palette.
}
\details{
BrBG
}
\examples{
BrBG(9)
ggPalette(BrBG(9))
}
