% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{BuGn}
\alias{BuGn}
\title{Create a BuGn palette}
\usage{
BuGn(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a BuGn palette.
}
\details{
BuGn
}
\examples{
BuGn(9)
ggPalette(BuGn(9))
}
