% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{BuPu}
\alias{BuPu}
\title{Create a BuPu palette}
\usage{
BuPu(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a BuPu palette.
}
\details{
BuPu
}
\examples{
BuPu(9)
ggPalette(BuPu(9))
}
