% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{cividis}
\alias{cividis}
\title{Create a cividis palette}
\usage{
cividis(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a cividis palette.
}
\details{
cividis
}
\examples{
cividis(9)
ggPalette(cividis(9))
}
