% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Col2Hsl.r
\name{Col2Hsl}
\alias{Col2Hsl}
\title{Col2Hsl}
\usage{
Col2Hsl(color.col = NULL, alpha.bln = FALSE)
}
\arguments{
\item{color.col}{<charcater>: a vector of the color's present in base R colors()}

\item{alpha.bln}{<logical>: a boolean that indicate if the alpha layer must be return. (Default FALSE)}
}
\value{
A vector of the color's hsl code
}
\description{
Convert a color in hsl (Hue,Saturation,Light) format
}
\examples{
Col2Hsl("#ebb0107f",alpha.bln=TRUE)
}
