% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Darken.r
\name{Darken}
\alias{Darken}
\title{Darken}
\usage{
Darken(color.col = NULL, alpha.bln = FALSE)
}
\arguments{
\item{color.col}{<character or numeric>: color in any format}

\item{alpha.bln}{<logical>: a boolean that indicate if the alpha layer must be return. (Default FALSE)}
}
\value{
An integer vector of the color's rgb code
}
\description{
Fixe light at 25%
}
\examples{
Darken("#ebb0107f",alpha.bln=TRUE)
}
