% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPalette.r
\name{ggPalette}
\alias{ggPalette}
\title{ggPalette}
\usage{
ggPalette(pal)
}
\arguments{
\item{pal}{<character or numeric>: color palette in Hexadecimal, R colour or RGB format}
}
\value{
a ggplot object
}
\description{
Create a ggplot object of a colour palette
}
\examples{
pal <- Hue(paletteLength.num=1,hueRange.num=c(0,120))
ggp <- ggPalette(pal)
file.pth <- DevToolKit::CheckFilePath(fileName.chr="Palette",format.chr="pdf" )
pdf(file=file.pth)
print(ggp)
dev.off()
}
