% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{GnBu}
\alias{GnBu}
\title{Create a GnBu palette}
\usage{
GnBu(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a GnBu palette.
}
\details{
GnBu
}
\examples{
GnBu(9)
ggPalette(GnBu(9))
}
