% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{Greens}
\alias{Greens}
\title{Create a Greens palette}
\usage{
Greens(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a Greens palette.
}
\details{
Greens
}
\examples{
Greens(9)
ggPalette(Greens(9))
}
