% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{Greys}
\alias{Greys}
\title{Create a Greys palette}
\usage{
Greys(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a Greys palette.
}
\details{
Greys
}
\examples{
Greys(9)
ggPalette(Greys(9))
}
