% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hex2Rgb.r
\name{Hex2Rgb}
\alias{Hex2Rgb}
\title{Hex2Rgb}
\usage{
Hex2Rgb(hex.col = NULL, alpha.bln = FALSE)
}
\arguments{
\item{hex.col}{<charcater>: a character of the color's hexadecimal code}

\item{alpha.bln}{<logical>: a boolean that indicate if the alpha layer must be return. (Default FALSE)}
}
\value{
An integer vector of the color's RGB code
}
\description{
Convert a color in hexadecimal format to RGB format
}
\examples{
Hex2Rgb("#ebb0107f",alpha.bln=TRUE)
}
