% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hsl2Hex.r
\name{Hsl2Hex}
\alias{Hsl2Hex}
\title{Hsl2Hex}
\usage{
Hsl2Hex(hsl.col = NULL, alpha.bln = FALSE)
}
\arguments{
\item{hsl.col}{<charcater>: a vector of the color's hsl code}

\item{alpha.bln}{<logical>: a boolean that indicate if the alpha layer must be return. (Default FALSE)}
}
\value{
A character of the color's hexadecimal code
}
\description{
Convert a color in hsl (Hue,Saturation,Light) format to hexadecimal format
}
\examples{
Hsl2Hex(c(43.8,0.873,0.492,0.498),alpha.bln=TRUE)
}
