% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hsl2Rgb.r
\name{Hsl2Rgb}
\alias{Hsl2Rgb}
\title{Hsl2Rgb}
\usage{
Hsl2Rgb(hsl.col = NULL, alpha.bln = FALSE)
}
\arguments{
\item{hsl.col}{<charcater>: a vector of the color's hsl code}

\item{alpha.bln}{<logical>: a boolean that indicate if the alpha layer must be return. (Default FALSE)}
}
\value{
An integer vector of the color's RGB code
}
\description{
Convert a color in hsl (Hue,Saturation,Light) format to RGB format
}
\examples{
Hsl2Rgb(c(43.8,0.873,0.492,0.498),alpha.bln=TRUE)
}
