% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{inferno}
\alias{inferno}
\title{Create a inferno palette}
\usage{
inferno(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a inferno palette.
}
\details{
inferno
}
\examples{
inferno(9)
ggPalette(inferno(9))
}
