% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.col.r
\name{is.col}
\alias{is.col}
\title{is.col}
\usage{
is.col(color.col = NULL)
}
\arguments{
\item{color.col}{<character or numeric>: a color}
}
\value{
Logical
}
\description{
Check if a color is in R colour format
}
\examples{
is.col("red")
is.col("#FFFFFF")
is.col(c(125,125,125))
is.col(c(43.8,0.873,0.492))
}
