% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.hex.r
\name{is.hex}
\alias{is.hex}
\title{is.hex}
\usage{
is.hex(color.col = NULL)
}
\arguments{
\item{color.col}{<character or numeric>: a color}
}
\value{
Logical
}
\description{
Check if a color is in hexadecimal colour format
}
\examples{
is.hex("red")
is.hex("#FFFFFF")
is.hex(c(125,125,125))
is.hex(c(43.8,0.873,0.492))
}
