% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsHex.r
\name{IsHex}
\alias{IsHex}
\title{IsHex}
\usage{
IsHex(color.col = NULL)
}
\arguments{
\item{color.col}{<character or numeric>: a color}
}
\value{
Logical
}
\description{
Check if a color is in hexadecimal colour format
}
\examples{
IsHex("red")
IsHex("#FFFFFF")
IsHex(c(125,125,125))
IsHex(c(43.8,0.873,0.492))
}
