% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsHsl.r
\name{IsHsl}
\alias{IsHsl}
\title{IsHsl}
\usage{
IsHsl(color.col = NULL)
}
\arguments{
\item{color.col}{<character or numeric>: a color}
}
\value{
Logical
}
\description{
Check if a color is in HSL colour format
}
\examples{
IsHsl("red")
IsHsl("#FFFFFF")
IsHsl(c(125,125,125))
IsHsl(c(43.8,0.873,0.492))
}
