% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsRgb.r
\name{IsRgb}
\alias{IsRgb}
\title{IsRgb}
\usage{
IsRgb(color.col = NULL)
}
\arguments{
\item{color.col}{<character or numeric>: a color}
}
\value{
Logical
}
\description{
Check if a color is in RGB colour format
}
\examples{
IsRgb("red")
IsRgb("#FFFFFF")
IsRgb(c(125,125,125))
IsRgb(c(43.8,0.873,0.492))
}
