% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{magma}
\alias{magma}
\title{Create a magma palette}
\usage{
magma(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a magma palette.
}
\details{
magma
}
\examples{
magma(9)
ggPalette(magma(9))
}
