% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{mako}
\alias{mako}
\title{Create a mako palette}
\usage{
mako(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a mako palette.
}
\details{
mako
}
\examples{
mako(9)
ggPalette(mako(9))
}
