% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{Oranges}
\alias{Oranges}
\title{Create a Oranges palette}
\usage{
Oranges(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a Oranges palette.
}
\details{
Oranges
}
\examples{
Oranges(9)
ggPalette(Oranges(9))
}
