% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{OrRd}
\alias{OrRd}
\title{Create a OrRd palette}
\usage{
OrRd(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a OrRd palette.
}
\details{
OrRd
}
\examples{
OrRd(9)
ggPalette(OrRd(9))
}
