% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{PiYg}
\alias{PiYg}
\title{Create a PiYg palette}
\usage{
PiYg(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a PiYg palette.
}
\details{
PiYg
}
\examples{
PiYg(9)
ggPalette(PiYg(9))
}
