% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{plasma}
\alias{plasma}
\title{Create a plasma palette}
\usage{
plasma(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a plasma palette.
}
\details{
plasma
}
\examples{
plasma(9)
ggPalette(plasma(9))
}
