% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPalette.r
\name{PlotPalette}
\alias{PlotPalette}
\title{PlotPalette}
\usage{
PlotPalette(pal, file.pth = NULL)
}
\arguments{
\item{pal}{<character or numeric>: color palette in Hexadecimal, R colour or RGB format}

\item{file.pth}{<character>: Path of pdf. If NULL, pdf is generated in work directory (Default NULL).}
}
\value{
None
}
\description{
Plot a colour palette
}
\examples{
pal <- Hue(paletteLength.num=1,hueRange.num=c(0,120))
PlotPalette(pal)
}
