% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{PrGn}
\alias{PrGn}
\title{Create a PrGn palette}
\usage{
PrGn(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a PrGn palette.
}
\details{
PrGn
}
\examples{
PrGn(9)
ggPalette(PrGn(9))
}
