% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{PuBu}
\alias{PuBu}
\title{Create a PuBu palette}
\usage{
PuBu(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a PuBu palette.
}
\details{
PuBu
}
\examples{
PuBu(9)
ggPalette(PuBu(9))
}
