% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{PuBuGn}
\alias{PuBuGn}
\title{Create a PuBuGn palette}
\usage{
PuBuGn(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a PuBuGn palette.
}
\details{
PuBuGn
}
\examples{
PuBuGn(9)
ggPalette(PuBuGn(9))
}
