% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{PuOr}
\alias{PuOr}
\title{Create a PuOr palette}
\usage{
PuOr(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a PuOr palette.
}
\details{
PuOr
}
\examples{
PuOr(9)
ggPalette(PuOr(9))
}
