% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{PuRd}
\alias{PuRd}
\title{Create a PuRd palette}
\usage{
PuRd(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a PuRd palette.
}
\details{
PuRd
}
\examples{
PuRd(9)
ggPalette(PuRd(9))
}
