% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{Purples}
\alias{Purples}
\title{Create a Purples palette}
\usage{
Purples(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a Purples palette.
}
\details{
Purples
}
\examples{
Purples(9)
ggPalette(Purples(9))
}
