% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{RdBu}
\alias{RdBu}
\title{Create a RdBu palette}
\usage{
RdBu(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a RdBu palette.
}
\details{
RdBu
}
\examples{
RdBu(9)
ggPalette(RdBu(9))
}
