% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{RdPu}
\alias{RdPu}
\title{Create a RdPu palette}
\usage{
RdPu(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a RdPu palette.
}
\details{
RdPu
}
\examples{
RdPu(9)
ggPalette(RdPu(9))
}
