% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{RdYlBu}
\alias{RdYlBu}
\title{Create a RdYlBu palette}
\usage{
RdYlBu(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a RdYlBu palette.
}
\details{
RdYlBu
}
\examples{
RdYlBu(9)
ggPalette(RdYlBu(9))
}
