% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{RdYlGn}
\alias{RdYlGn}
\title{Create a RdYlGn palette}
\usage{
RdYlGn(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a RdYlGn palette.
}
\details{
RdYlGn
}
\examples{
RdYlGn(9)
ggPalette(RdYlGn(9))
}
