% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{Reds}
\alias{Reds}
\title{Create a Reds palette}
\usage{
Reds(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a Reds palette.
}
\details{
Reds
}
\examples{
Reds(9)
ggPalette(Reds(9))
}
