% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgb2Hsl.r
\name{Rgb2Hsl}
\alias{Rgb2Hsl}
\title{Rgb2Hsl}
\usage{
Rgb2Hsl(rgb.col = NULL, alpha.bln = FALSE)
}
\arguments{
\item{rgb.col}{<integer>: an integer of the color's RGB code}

\item{alpha.bln}{<logical>: a boolean that indicate if the alpha layer must be return. (Default FALSE)}
}
\value{
A vector of the color's hsl code
}
\description{
Convert a color in RGB format to hsl (Hue,Saturation,Light) format
}
\examples{
Rgb2Hsl(c(235,176,16,127),alpha.bln=TRUE)
}
