% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{rocket}
\alias{rocket}
\title{Create a rocket palette}
\usage{
rocket(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a rocket palette.
}
\details{
rocket
}
\examples{
rocket(9)
ggPalette(rocket(9))
}
