% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{Spectral}
\alias{Spectral}
\title{Create a Spectral palette}
\usage{
Spectral(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a Spectral palette.
}
\details{
Spectral
}
\examples{
Spectral(9)
ggPalette(Spectral(9))
}
