% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{turbo}
\alias{turbo}
\title{Create a turbo palette}
\usage{
turbo(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a turbo palette.
}
\details{
turbo
}
\examples{
turbo(9)
ggPalette(turbo(9))
}
