% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{viridis}
\alias{viridis}
\title{Create a viridis palette}
\usage{
viridis(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a viridis palette.
}
\details{
viridis
}
\examples{
viridis(9)
ggPalette(viridis(9))
}
