% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{YlGn}
\alias{YlGn}
\title{Create a YlGn palette}
\usage{
YlGn(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a YlGn palette.
}
\details{
YlGn
}
\examples{
YlGn(9)
ggPalette(YlGn(9))
}
