% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{YlGnBu}
\alias{YlGnBu}
\title{Create a YlGnBu palette}
\usage{
YlGnBu(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a YlGnBu palette.
}
\details{
YlGnBu
}
\examples{
YlGnBu(9)
ggPalette(YlGnBu(9))
}
