% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{YlOrBr}
\alias{YlOrBr}
\title{Create a YlOrBr palette}
\usage{
YlOrBr(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a YlOrBr palette.
}
\details{
YlOrBr
}
\examples{
YlOrBr(9)
ggPalette(YlOrBr(9))
}
