% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.r
\name{YlOrRd}
\alias{YlOrRd}
\title{Create a YlOrRd palette}
\usage{
YlOrRd(n)
}
\arguments{
\item{n}{<numeric>: color number.}
}
\value{
A vector of color.
}
\description{
Create a YlOrRd palette.
}
\details{
YlOrRd
}
\examples{
YlOrRd(9)
ggPalette(YlOrRd(9))
}
