% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brighten.r
\name{Brighten}
\alias{Brighten}
\title{Fixe saturation at 100\% and light at 50\%.}
\usage{
Brighten(color.col = NULL, alpha.bln = FALSE)
}
\arguments{
\item{color.col}{<character or numeric>: color in any format.}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
An integer vector of the color's rgb code.
}
\description{
Fixe saturation at 100\% and light at 50\%.
}
\details{
Brighten
}
\examples{
Brighten("#ebb0107f",alpha.bln=TRUE)
}
