% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hex2Hsl.r
\name{Hex2Hsl}
\alias{Hex2Hsl}
\title{Hex2Hsl}
\usage{
Hex2Hsl(hex.col = NULL, alpha.bln = FALSE)
}
\arguments{
\item{hex.col}{<charcater>: a character of the color's hexadecimal code}

\item{alpha.bln}{<logical>: a boolean that indicate if the alpha layer must be return. (Default FALSE)}
}
\value{
A vector of the color's hsl code
}
\description{
Convert a color in hexadecimal format to hsl (Hue,Saturation,Light) format
}
\examples{
Hex2Hsl("#ebb0107f",alpha.bln=TRUE)
}
